requirements.txt
PyQt5>=5.15.10
pyinstaller>=6.6
openpyxl>=3.1.2
pymupdf>=1.24.0
python-dateutil>=2.9.0
pyautogui
python-docx
psycopg2-binary
arabic-reshaper
python-bidi
Flask
Flask-Login
Authlib
requests
gunicorn
python-dotenv
